@@ -20,22 +20,22 @@
   "dependencies": {
     "@google-cloud/functions-framework": "^3.3.0",
     "firebase-admin": "^12.0.0",
-    "firebase-functions": "^4.7.0",
-    "ufo": "^1.4.0"
+    "firebase-functions": "^4.8.1",
+    "ufo": "^1.5.3"
   },
   "devDependencies": {
     "@firebase/rules-unit-testing": "^3.0.1",
-    "@typescript-eslint/eslint-plugin": "^7.1.0",
-    "@typescript-eslint/parser": "^7.1.0",
+    "@typescript-eslint/eslint-plugin": "^7.3.1",
+    "@typescript-eslint/parser": "^7.3.1",
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-import": "^2.29.1",
     "firebase-functions-test": "^3.1.1",
-    "firebase-tools": "^13.4.0",
+    "firebase-tools": "^13.5.2",
     "npm-run-all": "^4.1.5",
     "prettier": "^3.2.5",
-    "typescript": "~5.3.3",
-    "vitest": "^1.3.1"
+    "typescript": "~5.4.2",
+    "vitest": "^1.4.0"
   },
   "private": true
 }
