@@ -1,3 +1,3 @@
-export const DEFAULT_FIREBASE_REGION = 'europe-west2' as const;
+export const DEFAULT_FIREBASE_REGION = 'europe-west2';
 
 export * from './models';
